+++ src/app/_components/_sharedcomponents/Preferences/PreferencesSubElementVariants/CurrentGameTab.tsx
@@ -17,6 +17,7 @@ function CurrentGameTab() {
     const router = useRouter();
     const currentPlayerName = gameState.players[connectedPlayer]?.name
     const [confirmConcede, setConfirmConcede] = useState<boolean>(false);
+    const [bugReportOpen, setBugReportOpen] = useState<boolean>(false);
 
     useEffect(() => {
         if(confirmConcede){
@@ -35,6 +36,16 @@ function CurrentGameTab() {
             // Reset the confirmation
             setConfirmConcede(false);
         }
+    };
+
+    // Handler for opening the bug report dialog
+    const handleOpenBugReport = () => {
+        setBugReportOpen(true);
+    };
+
+    // Handler for closing the bug report dialog
+    const handleCloseBugReport = () => {
+        setBugReportOpen(false);
     };
 
     // Handler for spectators to leave the game
@@ -111,7 +122,7 @@ function CurrentGameTab() {
                     Thank you for helping us test the beta!
                 </Typography>
                 <Typography sx={styles.typeographyStyle}>
-                    If you run indo any issues, please let us know in  
+                    If you run into any issues, please let us know in
                     <MuiLink
                         href="https://discord.gg/hKRaqHND4v"
                         target="_blank"
@@ -121,7 +132,18 @@ function CurrentGameTab() {
                         Discord
                     </MuiLink>. Thanks!
                 </Typography>
+                <Box sx={{ ...styles.contentContainer, mb:'20px' }}>
+                    <PreferenceButton variant={'standard'} text={'Report bug'} buttonFnc={handleOpenBugReport} />
+                    <Typography sx={styles.typeographyStyle}>
+                        Report a bug to the developer team.
+                    </Typography>
+                </Box>
             </Box>
+            {/* Bug Report Dialog */}
+            <BugReportDialog
+                open={bugReportOpen}
+                onClose={handleCloseBugReport}
+            />
         </>
     );
 }
